+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [1.10.0] - 2024-11-15
 ### Added
 * Added `MODEL_TYPE` option to `translateText()` to use models with higher
   translation quality (available for some language pairs), or better latency.
@@ -169,6 +169,7 @@ Stable release.
 Initial version.
 
 
+[1.10.0]: https://github.com/DeepLcom/deepl-php/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/DeepLcom/deepl-php/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/DeepLcom/deepl-php/compare/v1.7.2...v1.8.0
 [1.7.2]: https://github.com/DeepLcom/deepl-php/compare/v1.7.1...v1.7.2